Add Gemfile. Use Twitter gem. Update script.
Was previously using Twitter4R gem. Updating to use Twitter gem
as this is now more actively maintained. For the most part, the
API is consistent with Twitter4R, so miniminal changes were needed
to make the script function, but there were a couple of differences
that needed to be accounted for. To follow what I believe to be
best practices for gem use, a bundler Gemfile was added. For more
information, see gembundler.com.