Use batching in requests.
This takes advantage of the new Twitter gem's better support for
batching. It also adds a command line option for the script to specify
the batch size, but be default it will use 100, which is the maximum.
(The command line option could be removed...I'm not sure why someone
would want to specify a lower number, but if that use case exists it is
possible right now.) Some updates to keep track of how many screen names
do not exist had to be done (basically, we can't just increase the
not_found variable in the caught exception.)